+++ src/components/mypage/tabs/MyReview.tsx
@@ -4,7 +4,7 @@ import CardLayout from '../common/CardLayout';
 const MyReview = () => {
   return (
     <CardLayout title='리뷰 관리'>
-
+      <h4 className='py-2 border-b border-solid border-[#abc8df] text-[#1B304A] font-semibold'>내 리뷰 (0)</h4>
     </CardLayout>
   );
 };
